Term collections renamed to utilize Underscore naming scheme
- All Importers and collection schema definitions updated to utilize underscore _XXX term naming scheme instead of termRarity etc.
