+++ app/_custom/collections/lightCone.ts
@@ -48,15 +48,15 @@ export const LightCone: CollectionConfig = {
          type: "text",
       },
       {
-         name: "term_rarity",
+         name: "rarity",
          type: "relationship",
-         relationTo: "termRarity-lKJ16E5IhH",
+         relationTo: "_rarity-lKJ16E5IhH",
          hasMany: false,
       },
       {
-         name: "term_path",
+         name: "path",
          type: "relationship",
-         relationTo: "termPath-lKJ16E5IhH",
+         relationTo: "_path-lKJ16E5IhH",
          hasMany: false,
       },
       {
