+++ app/_custom/collections/materials.ts
@@ -46,16 +46,16 @@ export const Materials: CollectionConfig = {
          type: "text",
       },
       {
-         name: "term_itemtype",
+         name: "itemtype",
          type: "relationship",
-         relationTo: "termItemtype-lKJ16E5IhH",
+         relationTo: "_itemType-lKJ16E5IhH",
          hasMany: false,
          required: false,
       },
       {
-         name: "term_rarity",
+         name: "rarity",
          type: "relationship",
-         relationTo: "termRarity-lKJ16E5IhH",
+         relationTo: "_rarity-lKJ16E5IhH",
          hasMany: false,
          required: false,
       },
