+++ app/_custom/import/import_201_LightCone.ts
@@ -4,10 +4,10 @@ require("dotenv").config();
 
 const { PAYLOADCMS_SECRET, MONGO_URL } = process.env;
 
-//Array of objects matching the payload shape, change to match your needs
-const collectionName = "character";
-const data = require("./import_files/Character.json");
-const idField = "character_id";
+//Array of objects matching the payload shape, change to match your need
+const collectionName = "lightCone";
+const data = require("./import_files/LightCone.json");
+const idField = "lightcone_id";
 const siteId = "lKJ16E5IhH";
 const userId = "63fec4372464d0e4c5c316e7"; // NorseFTX@gamepress.gg User ID for author field
 
@@ -45,26 +45,14 @@ const seedUploads = async (result: any) => {
 
 	// Define Image fields (global)
 	const iconImport = {
-		image_round_icon: siteId + "_" + result.image_round_icon?.name.replace(".png",""),
-		image_action: siteId + "_" + result.image_action?.name.replace(".png",""),
-		image_battle_detail: siteId + "_" + result.image_battle_detail?.name.replace(".png",""),
 		image_full: siteId + "_" + result.image_full?.name.replace(".png",""),
-		image_full_bg: siteId + "_" + result.image_full_bg?.name.replace(".png",""),
-		image_full_front: siteId + "_" + result.image_full_front?.name.replace(".png",""),
 	}
-	
+
 	// Unlock Materials array
 	var matData: any = [];
-	var rewardListImport: any = [];
-	var rewardMaxListImport: any = [];
 	var promotionCostImport: any = [];
 
-	// Get unique list of Materials in full array
-	// .split("data_key\":") -> Separate array by data key value
-	// .map(a => a.replace(/}.*/,"")) -> Removes extraneous text after value
-	// .slice(1) // Removes first element of array that has extraneous text
-	// .filter((v,i,a) => a.indexOf(v) === i) -> Gets unique values
-	const allmatjson = JSON.stringify(result.reward_list) + JSON.stringify(result.reward_max_list) + JSON.stringify(result.promotion_cost);
+	const allmatjson = JSON.stringify(result.promotion_cost);
 	const matList = allmatjson?.split("data_key\":").map(a => a.replace(/}.*/,"")).slice(1).filter((v,i,a) => a.indexOf(v) === i);
 
 	if (matList?.length > 0) {
@@ -89,25 +77,6 @@ const seedUploads = async (result: any) => {
 		}));
 	}
 
-
-	if (result.reward_list.length > 0) {
-		rewardListImport = result.reward_list.map((l:any) => {
-			const matId = matData.find((a:any) => a.data_key == l.materials?.data_key)?.id;
-			return {
-				...l,
-				materials: matId,
-			}
-		});
-	}
-	if (result.reward_max_list.length > 0) {
-		rewardMaxListImport = result.reward_max_list.map((l:any) => {
-			const matId = matData.find((a:any) => a.data_key == l.materials?.data_key)?.id;
-			return {
-				...l,
-				materials: matId,
-			}
-		});
-	}
 	if (result.promotion_cost.length > 0) {
 		promotionCostImport = result.promotion_cost.map((l:any) => {
 			const matQty = l.material_qty?.map((mat:any) => {
@@ -127,8 +96,6 @@ const seedUploads = async (result: any) => {
 	}
 
 	var matQtyImport = {
-		reward_list: rewardListImport,
-		reward_max_list: rewardMaxListImport,
 		promotion_cost: promotionCostImport,
 	}
 
@@ -139,17 +106,15 @@ const seedUploads = async (result: any) => {
 
 	var relationFields: any = {};
 	// Single relation fields
-	relationFields["term_rarity"] = null;
-	relationFields["term_element"] = null;
-	relationFields["term_path"] = null;
+	relationFields["rarity"] = null;
+	relationFields["path"] = null;
 
 	// Multi relation fields
-	relationFields["eidolons"] = null;
-	relationFields["traces"] = null;
+	// relationFields["skill_data"] = null;
 
-	var fieldName = "term_rarity";
+	var fieldName = "rarity";
 	var idName = "name";
-	var collName = "termRarity";
+	var collName = "_rarity";
 	if (result[fieldName]?.[idName]) {
 		const relEntry = await payload.find({
 			collection: collName + "-" + siteId,
@@ -164,26 +129,9 @@ const seedUploads = async (result: any) => {
 			relationFields[fieldName] = relEntry?.docs?.[0]?.id;
 		}
 	}
-	var fieldName = "term_element";
+	var fieldName = "path";
 	var idName = "data_key";
-	var collName = "termElement";
-	if (result[fieldName]?.[idName]) {
-		const relEntry = await payload.find({
-			collection: collName + "-" + siteId,
-			where: {
-				[idName]: {
-					equals: result[fieldName]?.[idName],
-				},
-			}
-		});
-
-		if (relEntry?.docs?.[0]?.id) {
-			relationFields[fieldName] = relEntry?.docs?.[0]?.id;
-		}
-	}
-	var fieldName = "term_path";
-	var idName = "data_key";
-	var collName = "termPath";
+	var collName = "_path";
 	if (result[fieldName]?.[idName]) {
 		const relEntry = await payload.find({
 			collection: collName + "-" + siteId,
@@ -199,57 +147,35 @@ const seedUploads = async (result: any) => {
 		}
 	}
 
-	if (result.traces?.length > 0) {
-		const traceEntry = await Promise.all(result.traces.map(async (t:any) => {
-			const findTrace = await payload.find({
-				collection: "trace-" + siteId,
-				where: {
-					trace_id: {
-						equals: t.trace_id,
-					},
-				}
-			});
-			if (findTrace?.docs?.[0]?.id) {
-				return findTrace?.docs?.[0]?.id;
-			}
-			else {
-				return null;
-			}
-	}));
+	// if (result.traces?.length > 0) {
+	// 	const traceEntry = await Promise.all(result.traces.map(async (t:any) => {
+	// 		const findTrace = await payload.find({
+	// 			collection: "trace-" + siteId,
+	// 			where: {
+	// 				trace_id: {
+	// 					equals: t.trace_id,
+	// 				},
+	// 			}
+	// 		});
+	// 		if (findTrace?.docs?.[0]?.id) {
+	// 			return findTrace?.docs?.[0]?.id;
+	// 		}
+	// 		else {
+	// 			return null;
+	// 		}
+	// }));
 
-		if (traceEntry.length > 0) {
-			relationFields["traces"] = traceEntry;
-		}
-	}
-
-	if (result.eidolons?.length > 0) {
-		const traceEntry = await Promise.all(result.eidolons.map(async (t:any) => {
-			const findTrace = await payload.find({
-				collection: "eidolon-" + siteId,
-				where: {
-					eidolon_id: {
-						equals: t.eidolon_id,
-					},
-				}
-			});
-			if (findTrace?.docs?.[0]?.id) {
-				return findTrace?.docs?.[0]?.id;
-			}
-			else {
-				return null;
-			}
-	}));
-
-		if (traceEntry.length > 0) {
-			relationFields["eidolons"] = traceEntry;
-		}
-	}
+	// 	if (traceEntry.length > 0) {
+	// 		relationFields["traces"] = traceEntry;
+	// 	}
+	// }
 
 	// ====================================
 	// End of Relation Fields
 	// ====================================
 	// ====================================
-	
+
+
 	// Check if entry exists
 	const existingEntry = await payload.find({
 		collection: collectionName + "-" + siteId,
@@ -279,7 +205,7 @@ const seedUploads = async (result: any) => {
 			id: baseID,
 			data: baseData,
 		});
-		sleep(50);
+		sleep(200);
 		console.log(`${JSON.stringify(updateItem)} Entry updated!`);
 
 		const itemId = updateItem.id;
@@ -288,9 +214,9 @@ const seedUploads = async (result: any) => {
 			...result,
 			entry: itemId,
 			id: collectionName + "-" + itemId,
+			...iconImport,
 			...matQtyImport,
 			...relationFields,
-			...iconImport,
 		};
 
 		const updateItemCustom = await payload.update({
@@ -315,7 +241,7 @@ const seedUploads = async (result: any) => {
 			data: baseData,
 		});
 		//Limit speed
-		sleep(50);
+		sleep(200);
 		console.log(`${JSON.stringify(createItem)} Import completed!`);
 		
 		const itemId = createItem.id;
@@ -324,9 +250,9 @@ const seedUploads = async (result: any) => {
 			...result,
 			entry: itemId,
 			id: collectionName + "-" + itemId,
+			...iconImport,
 			...matQtyImport,
 			...relationFields,
-			...iconImport,
 		};
 
 		const createItemCustom = await payload.create({
