+++ app/_custom/import/import_300c__Element.ts
@@ -4,9 +4,9 @@ require("dotenv").config();
 
 const { PAYLOADCMS_SECRET, MONGO_URL } = process.env;
 
-//Array of objects matching the payload shape, change to match your need
-const collectionName = "termStattype";
-const data = require("./import_files/" + collectionName + ".json");
+//Array of objects matching the payload shape, change to match your needs
+const collectionName = "_element";
+const data = require("./import_files/" + "_Element" + ".json");
 const idField = "data_key";
 const siteId = "lKJ16E5IhH";
 const userId = "63fec4372464d0e4c5c316e7"; // NorseFTX@gamepress.gg User ID for author field
@@ -42,17 +42,8 @@ const getData = async () =>
 const seedUploads = async (result: any) => {
 
 	const idValue = result[idField];
-
-	// Define Image fields (global)
-	const iconImport = {
-		icon_color: siteId + "_" + result.icon_color?.name.replace(".png",""),
-		icon_active: siteId + "_" + result.icon_active?.name.replace(".png",""),
-		icon_inactive: siteId + "_" + result.icon_inactive?.name.replace(".png",""),
-		icon_damage_res: siteId + "_" + result.icon_damage_res?.name.replace(".png",""),
-		icon_small: siteId + "_" + result.icon_small?.name.replace(".png",""),
-	}
-
 	// Check if entry exists
+	
 	const existingEntry = await payload.find({
 		collection: collectionName + "-" + siteId,
 		where: {
@@ -81,7 +72,7 @@ const seedUploads = async (result: any) => {
 			id: baseID,
 			data: baseData,
 		});
-		sleep(50);
+		sleep(200);
 		console.log(`${JSON.stringify(updateItem)} Entry updated!`);
 
 		const itemId = updateItem.id;
@@ -90,7 +81,10 @@ const seedUploads = async (result: any) => {
 			...result,
 			entry: itemId,
 			id: collectionName + "-" + itemId,
-			...iconImport,
+			icon_color: siteId + "_" + result.icon_color?.name.replace(".png",""),
+			icon_active: siteId + "_" + result.icon_active?.name.replace(".png",""),
+			icon_inactive: siteId + "_" + result.icon_inactive?.name.replace(".png",""),
+			icon_damage_res: siteId + "_" + result.icon_damage_res?.name.replace(".png",""),
 		};
 
 		const updateItemCustom = await payload.update({
@@ -115,7 +109,7 @@ const seedUploads = async (result: any) => {
 			data: baseData,
 		});
 		//Limit speed
-		sleep(50);
+		sleep(200);
 		console.log(`${JSON.stringify(createItem)} Import completed!`);
 		
 		const itemId = createItem.id;
@@ -124,7 +118,10 @@ const seedUploads = async (result: any) => {
 			...result,
 			entry: itemId,
 			id: collectionName + "-" + itemId,
-			...iconImport,
+			icon_color: siteId + "_" + result.icon_color?.name.replace(".png",""),
+			icon_active: siteId + "_" + result.icon_active?.name.replace(".png",""),
+			icon_inactive: siteId + "_" + result.icon_inactive?.name.replace(".png",""),
+			icon_damage_res: siteId + "_" + result.icon_damage_res?.name.replace(".png",""),
 		};
 
 		const createItemCustom = await payload.create({
