--- app/_custom/import/import_files/Eidolon.json
@@ -1,122 +1,0 @@
-[
-	{
-		"eidolon_id": 100101,
-		"name": "Memory of You",
-		"description": "Increases the number of times the Talent's Counter effect can be triggered each turn by 1. When Counter is triggered, deals additional Ice DMG equal to <unbreak>60%</unbreak> of March 7th's DEF.",
-		"rank": 1,
-		"icon_name": "SkillIcon_1001_Rank1",
-		"trigger": 2089636447,
-		"unlock_materials": [
-			{
-				"materials": {
-					"data_key": 11001
-				},
-				"qty": 1
-			}
-		],
-		"icon": {
-			"name": "SkillIcon_1001_Rank1.png"
-		},
-		"checksum": "12a35753"
-	},
-	{
-		"eidolon_id": 100102,
-		"name": "Memory of It",
-		"description": "Every time March 7th's Ultimate Freezes a target, she regenerates <unbreak>20</unbreak> Energy.",
-		"rank": 2,
-		"icon_name": "SkillIcon_1001_Rank2",
-		"trigger": 523552506,
-		"unlock_materials": [
-			{
-				"materials": {
-					"data_key": 11001
-				},
-				"qty": 1
-			}
-		],
-		"icon": {
-			"name": "SkillIcon_1001_Rank2.png"
-		},
-		"checksum": "1273ad9e"
-	},
-	{
-		"eidolon_id": 100103,
-		"name": "Memory of Everything",
-		"description": "Ultimate Lv. +2, up to a maximum of Lv. <unbreak>15</unbreak>. Basic ATK Lv. +1, up to a maximum of Lv. <unbreak>10</unbreak>.",
-		"rank": 3,
-		"icon_name": "SkillIcon_1001_Ultra",
-		"trigger": -1042531435,
-		"unlock_materials": [
-			{
-				"materials": {
-					"data_key": 11001
-				},
-				"qty": 1
-			}
-		],
-		"icon": {
-			"name": "SkillIcon_1001_Ultra.png"
-		},
-		"checksum": "128556cc"
-	},
-	{
-		"eidolon_id": 100104,
-		"name": "Don't Want to Lose",
-		"description": "Upon entering battle, grants a Shield equal to <unbreak>100%</unbreak> of March 7th's DEF plus <unbreak>400</unbreak> to the ally with the lowest HP percentage, lasting for <unbreak>6</unbreak> turn(s).",
-		"rank": 4,
-		"icon_name": "SkillIcon_1001_Rank4",
-		"trigger": 1686351920,
-		"unlock_materials": [
-			{
-				"materials": {
-					"data_key": 11001
-				},
-				"qty": 1
-			}
-		],
-		"icon": {
-			"name": "SkillIcon_1001_Rank4.png"
-		},
-		"checksum": "12aad07c"
-	},
-	{
-		"eidolon_id": 100105,
-		"name": "Don't Want to Forget",
-		"description": "Skill Lv. +2, up to a maximum of Lv. <unbreak>15</unbreak>. Talent Lv. +2, up to a maximum of Lv. <unbreak>15</unbreak>.",
-		"rank": 5,
-		"icon_name": "SkillIcon_1001_BP",
-		"trigger": 120267979,
-		"unlock_materials": [
-			{
-				"materials": {
-					"data_key": 11001
-				},
-				"qty": 1
-			}
-		],
-		"icon": {
-			"name": "SkillIcon_1001_BP.png"
-		},
-		"checksum": "127e3343"
-	},
-	{
-		"eidolon_id": 100106,
-		"name": "Just Like This, Always...",
-		"description": "Allies under the protection of the Shield granted by the Skill restores HP equal to <unbreak>12%</unbreak> of their Max HP plus <unbreak>96</unbreak> at the beginning of each turn.",
-		"rank": 6,
-		"icon_name": "SkillIcon_1001_Rank6",
-		"trigger": -1445815962,
-		"unlock_materials": [
-			{
-				"materials": {
-					"data_key": 11001
-				},
-				"qty": 1
-			}
-		],
-		"icon": {
-			"name": "SkillIcon_1001_Rank6.png"
-		},
-		"checksum": "12a2b660"
-	}
-]
