--- app/_custom/import/import_files/termElement.json
@@ -1,1 +1,0 @@
-[{"data_key":"Physical","icon_path":"SpriteOutput/UI/Nature/IconAttribute/IconAttributePhysical.png","icon_color_path":"SpriteOutput/UI/Nature/IconNatureColor/IconNaturePhysical.png","icon_active_path":"SpriteOutput/UI/Nature/IconNatureForWeak/IconNatureWhitePhysical_2.png","icon_inactive_path":"SpriteOutput/UI/Nature/IconNatureForWeak/IconNatureWhitePhysical_1.png","icon_damage_res_path":"SpriteOutput/UI/Avatar/Icon/IconPhysicalResistanceDelta.png","icon_name":"IconAttributePhysical","icon_color_name":"IconNaturePhysical","icon_active_name":"IconNatureWhitePhysical_2","icon_inactive_name":"IconNatureWhitePhysical_1","icon_damage_res_name":"IconPhysicalResistanceDelta","name":"Physical","icon":{"name":"IconAttributePhysical.png"},"icon_color":{"name":"IconNaturePhysical.png"},"icon_active":{"name":"IconNatureWhitePhysical_2.png"},"icon_inactive":{"name":"IconNatureWhitePhysical_1.png"},"icon_damage_res":{"name":"IconPhysicalResistanceDelta.png"},"checksum":"14af654b"},{"data_key":"Fire","icon_path":"SpriteOutput/UI/Nature/IconAttribute/IconAttributeFire.png","icon_color_path":"SpriteOutput/UI/Nature/IconNatureColor/IconNatureColorFire.png","icon_active_path":"SpriteOutput/UI/Nature/IconNatureForWeak/IconNatureWhiteFire_2.png","icon_inactive_path":"SpriteOutput/UI/Nature/IconNatureForWeak/IconNatureWhiteFire_1.png","icon_damage_res_path":"SpriteOutput/UI/Avatar/Icon/IconFireResistanceDelta.png","icon_name":"IconAttributeFire","icon_color_name":"IconNatureColorFire","icon_active_name":"IconNatureWhiteFire_2","icon_inactive_name":"IconNatureWhiteFire_1","icon_damage_res_name":"IconFireResistanceDelta","name":"Fire","icon":{"name":"IconAttributeFire.png"},"icon_color":{"name":"IconNatureColorFire.png"},"icon_active":{"name":"IconNatureWhiteFire_2.png"},"icon_inactive":{"name":"IconNatureWhiteFire_1.png"},"icon_damage_res":{"name":"IconFireResistanceDelta.png"},"checksum":"14636cc7"},{"data_key":"Ice","icon_path":"SpriteOutput/UI/Nature/IconAttribute/IconAttributeIce.png","icon_color_path":"SpriteOutput/UI/Nature/IconNatureColor/IconNatureColorIce.png","icon_active_path":"SpriteOutput/UI/Nature/IconNatureForWeak/IconNatureWhiteIce_2.png","icon_inactive_path":"SpriteOutput/UI/Nature/IconNatureForWeak/IconNatureWhiteIce_1.png","icon_damage_res_path":"SpriteOutput/UI/Avatar/Icon/IconIceResistanceDelta.png","icon_name":"IconAttributeIce","icon_color_name":"IconNatureColorIce","icon_active_name":"IconNatureWhiteIce_2","icon_inactive_name":"IconNatureWhiteIce_1","icon_damage_res_name":"IconIceResistanceDelta","name":"Ice","icon":{"name":"IconAttributeIce.png"},"icon_color":{"name":"IconNatureColorIce.png"},"icon_active":{"name":"IconNatureWhiteIce_2.png"},"icon_inactive":{"name":"IconNatureWhiteIce_1.png"},"icon_damage_res":{"name":"IconIceResistanceDelta.png"},"checksum":"144b853e"},{"data_key":"Thunder","icon_path":"SpriteOutput/UI/Nature/IconAttribute/IconAttributeThunder.png","icon_color_path":"SpriteOutput/UI/Nature/IconNatureColor/IconNatureColorThunder.png","icon_active_path":"SpriteOutput/UI/Nature/IconNatureForWeak/IconNatureWhiteThunder_2.png","icon_inactive_path":"SpriteOutput/UI/Nature/IconNatureForWeak/IconNatureWhiteThunder_1.png","icon_damage_res_path":"SpriteOutput/UI/Avatar/Icon/IconThunderResistanceDelta.png","icon_name":"IconAttributeThunder","icon_color_name":"IconNatureColorThunder","icon_active_name":"IconNatureWhiteThunder_2","icon_inactive_name":"IconNatureWhiteThunder_1","icon_damage_res_name":"IconThunderResistanceDelta","name":"Lightning","icon":{"name":"IconAttributeThunder.png"},"icon_color":{"name":"IconNatureColorThunder.png"},"icon_active":{"name":"IconNatureWhiteThunder_2.png"},"icon_inactive":{"name":"IconNatureWhiteThunder_1.png"},"icon_damage_res":{"name":"IconThunderResistanceDelta.png"},"checksum":"14afb7e6"},{"data_key":"Wind","icon_path":"SpriteOutput/UI/Nature/IconAttribute/IconAttributeWind.png","icon_color_path":"SpriteOutput/UI/Nature/IconNatureColor/IconNatureColorWind.png","icon_active_path":"SpriteOutput/UI/Nature/IconNatureForWeak/IconNatureWhiteWind_2.png","icon_inactive_path":"SpriteOutput/UI/Nature/IconNatureForWeak/IconNatureWhiteWind_1.png","icon_damage_res_path":"SpriteOutput/UI/Avatar/Icon/IconWindResistanceDelta.png","icon_name":"IconAttributeWind","icon_color_name":"IconNatureColorWind","icon_active_name":"IconNatureWhiteWind_2","icon_inactive_name":"IconNatureWhiteWind_1","icon_damage_res_name":"IconWindResistanceDelta","name":"Wind","icon":{"name":"IconAttributeWind.png"},"icon_color":{"name":"IconNatureColorWind.png"},"icon_active":{"name":"IconNatureWhiteWind_2.png"},"icon_inactive":{"name":"IconNatureWhiteWind_1.png"},"icon_damage_res":{"name":"IconWindResistanceDelta.png"},"checksum":"1464cb24"},{"data_key":"Quantum","icon_path":"SpriteOutput/UI/Nature/IconAttribute/IconAttributeQuantum.png","icon_color_path":"SpriteOutput/UI/Nature/IconNatureColor/IconNatureColorQuantum.png","icon_active_path":"SpriteOutput/UI/Nature/IconNatureForWeak/IconNatureWhiteQuantum_2.png","icon_inactive_path":"SpriteOutput/UI/Nature/IconNatureForWeak/IconNatureWhiteQuantum_1.png","icon_damage_res_path":"SpriteOutput/UI/Avatar/Icon/IconQuantumResistanceDelta.png","icon_name":"IconAttributeQuantum","icon_color_name":"IconNatureColorQuantum","icon_active_name":"IconNatureWhiteQuantum_2","icon_inactive_name":"IconNatureWhiteQuantum_1","icon_damage_res_name":"IconQuantumResistanceDelta","name":"Quantum","icon":{"name":"IconAttributeQuantum.png"},"icon_color":{"name":"IconNatureColorQuantum.png"},"icon_active":{"name":"IconNatureWhiteQuantum_2.png"},"icon_inactive":{"name":"IconNatureWhiteQuantum_1.png"},"icon_damage_res":{"name":"IconQuantumResistanceDelta.png"},"checksum":"14aef50d"},{"data_key":"Imaginary","icon_path":"SpriteOutput/UI/Nature/IconAttribute/IconAttributeImaginary.png","icon_color_path":"SpriteOutput/UI/Nature/IconNatureColor/IconNatureColorImaginary.png","icon_active_path":"SpriteOutput/UI/Nature/IconNatureForWeak/IconNatureWhiteImaginary_2.png","icon_inactive_path":"SpriteOutput/UI/Nature/IconNatureForWeak/IconNatureWhiteImaginary_1.png","icon_damage_res_path":"SpriteOutput/UI/Avatar/Icon/IconImaginaryResistanceDelta.png","icon_name":"IconAttributeImaginary","icon_color_name":"IconNatureColorImaginary","icon_active_name":"IconNatureWhiteImaginary_2","icon_inactive_name":"IconNatureWhiteImaginary_1","icon_damage_res_name":"IconImaginaryResistanceDelta","name":"Imaginary","icon":{"name":"IconAttributeImaginary.png"},"icon_color":{"name":"IconNatureColorImaginary.png"},"icon_active":{"name":"IconNatureWhiteImaginary_2.png"},"icon_inactive":{"name":"IconNatureWhiteImaginary_1.png"},"icon_damage_res":{"name":"IconImaginaryResistanceDelta.png"},"checksum":"14dcea9b"}]
