--- app/_custom/import/import_files/termItemtype.json
@@ -1,1 +1,0 @@
-[{"data_key":101,"inv_type":"Normal","sort_weight":10,"type_categories":["Virtual","Placeholder","Material"],"icon_path":"SpriteOutput/TabIcon/Inventory/InventoryFosterIcon.png","icon_name":"InventoryFosterIcon","name":"Upgrade Materials","icon":{"name":"InventoryFosterIcon.png"},"checksum":"1266e8e5"},{"data_key":102,"inv_type":"Normal","sort_weight":9,"type_categories":["Equipment"],"icon_path":"SpriteOutput/TabIcon/Inventory/InventoryLightConeIcon.png","icon_name":"InventoryLightConeIcon","name":"Light Cones","icon":{"name":"InventoryLightConeIcon.png"},"checksum":"125f915c"},{"data_key":103,"inv_type":"Normal","sort_weight":5,"type_categories":["Mission"],"icon_path":"SpriteOutput/TabIcon/Inventory/InventoryQuestIcon.png","icon_name":"InventoryQuestIcon","name":"Missions","icon":{"name":"InventoryQuestIcon.png"},"checksum":"1259fd63"},{"data_key":104,"inv_type":"Normal","sort_weight":6,"type_categories":["Food"],"icon_path":"SpriteOutput/TabIcon/Inventory/InventoryConsumablesIcon.png","icon_name":"InventoryConsumablesIcon","name":"Consumables","icon":{"name":"InventoryConsumablesIcon.png"},"checksum":"1260447d"},{"data_key":105,"inv_type":"Normal","sort_weight":4,"type_categories":["Gift","MusicAlbum","ValuableMaterial"],"icon_path":"SpriteOutput/TabIcon/Inventory/InventoryOtherIcon.png","icon_name":"InventoryOtherIcon","name":"Valuables","icon":{"name":"InventoryOtherIcon.png"},"checksum":"1263fac8"},{"data_key":106,"inv_type":"Normal","sort_weight":8,"type_categories":["Relic"],"icon_path":"SpriteOutput/TabIcon/Inventory/InventoryIcon.png","icon_name":"InventoryIcon","name":"Relics","icon":{"name":"InventoryIcon.png"},"checksum":"125378a6"},{"data_key":107,"inv_type":"Normal","sort_weight":7,"type_categories":["NonEnhancementMaterial"],"icon_path":"SpriteOutput/TabIcon/Inventory/InventoryUnfosterIcon.png","icon_name":"InventoryUnfosterIcon","name":"Other Materials","icon":{"name":"InventoryUnfosterIcon.png"},"checksum":"12651f14"}]
