--- app/_custom/import/import_files/termPath.json
@@ -1,1 +1,0 @@
-[{"data_key":"Warrior","icon_path":"SpriteOutput/AvatarProfessionTattoo/Profession/BgPathsWarrior.png","icon_small_path":"SpriteOutput/ProfessionIconSmall/IconProfessionWarriorSmall.png","icon_name":"BgPathsWarrior","icon_small_name":"IconProfessionWarriorSmall","name":"The Destruction","description":"Deal outstanding amounts of <span style=\"\"color: #f29e38ff\"\">damage</span> and possess great <span style=\"\"color: #f29e38ff\"\">survivability</span>. Suitable for various combat scenarios","icon":{"name":"BgPathsWarrior.png"},"icon_small":{"name":"IconProfessionWarriorSmall.png"},"checksum":"131ba57e"},{"data_key":"Rogue","icon_path":"SpriteOutput/AvatarProfessionTattoo/Profession/BgPathsRogue.png","icon_small_path":"SpriteOutput/ProfessionIconSmall/IconProfessionRogueSmall.png","icon_name":"BgPathsRogue","icon_small_name":"IconProfessionRogueSmall","name":"The Hunt","description":"Deal extraordinary amounts of <span style=\"\"color: #f29e38ff\"\">single-target damage</span>. The main damagers against Elite Enemies","icon":{"name":"BgPathsRogue.png"},"icon_small":{"name":"IconProfessionRogueSmall.png"},"checksum":"12e0c749"},{"data_key":"Mage","icon_path":"SpriteOutput/AvatarProfessionTattoo/Profession/BgPathsnMage.png","icon_small_path":"SpriteOutput/ProfessionIconSmall/IconProfessionMageSmall.png","icon_name":"BgPathsnMage","icon_small_name":"IconProfessionMageSmall","name":"The Erudition","description":"Deal remarkable amounts of <span style=\"\"color: #f29e38ff\"\">multi-target damage</span>. The main damagers against groups of enemies","icon":{"name":"BgPathsnMage.png"},"icon_small":{"name":"IconProfessionMageSmall.png"},"checksum":"12e0fff5"},{"data_key":"Shaman","icon_path":"SpriteOutput/AvatarProfessionTattoo/Profession/BgPathsShaman.png","icon_small_path":"SpriteOutput/ProfessionIconSmall/IconProfessionShamanSmall.png","icon_name":"BgPathsShaman","icon_small_name":"IconProfessionShamanSmall","name":"The Harmony","description":"Apply <span style=\"\"color: #f29e38ff\"\">buffs</span> to allies to improve the team's combat capacities","icon":{"name":"BgPathsShaman.png"},"icon_small":{"name":"IconProfessionShamanSmall.png"},"checksum":"12d322d2"},{"data_key":"Warlock","icon_path":"SpriteOutput/AvatarProfessionTattoo/Profession/BgPathsWarlock.png","icon_small_path":"SpriteOutput/ProfessionIconSmall/IconProfessionWarlockSmall.png","icon_name":"BgPathsWarlock","icon_small_name":"IconProfessionWarlockSmall","name":"The Nihility","description":"Apply <span style=\"\"color: #f29e38ff\"\">debuffs</span> to enemies to reduce their combat capacities","icon":{"name":"BgPathsWarlock.png"},"icon_small":{"name":"IconProfessionWarlockSmall.png"},"checksum":"12d77475"},{"data_key":"Knight","icon_path":"SpriteOutput/AvatarProfessionTattoo/Profession/BgPathsKnight.png","icon_small_path":"SpriteOutput/ProfessionIconSmall/IconProfessionKnightSmall.png","icon_name":"BgPathsKnight","icon_small_name":"IconProfessionKnightSmall","name":"The Preservation","description":"Possess powerful <span style=\"\"color: #f29e38ff\"\">defensive</span> abilities to protect allies in various ways","icon":{"name":"BgPathsKnight.png"},"icon_small":{"name":"IconProfessionKnightSmall.png"},"checksum":"12dddad1"},{"data_key":"Priest","icon_path":"SpriteOutput/AvatarProfessionTattoo/Profession/BgPathsPirest.png","icon_small_path":"SpriteOutput/ProfessionIconSmall/IconProfessionPirestSmall.png","icon_name":"BgPathsPirest","icon_small_name":"IconProfessionPirestSmall","name":"The Abundance","description":"<span style=\"\"color: #f29e38ff\"\">Heal</span> allies and restore HP to the team","icon":{"name":"BgPathsPirest.png"},"icon_small":{"name":"IconProfessionPirestSmall.png"},"checksum":"12c57ecb"}]
