+++ app/_custom/seed/example.json
@@ -1,24 +1,24 @@
 [
-{ "id": "AchievementBadge01"},
-{ "id": "AchievementBadge02"},
-{ "id": "AchievementBadge03"},
-{ "id": "BattleAchievementIcon"},
-{ "id": "BattleAchievementIcon_s"},
-{ "id": "ChallengeAchievementIcon"},
-{ "id": "ChallengeAchievementIcon_s"},
-{ "id": "CollectAchievementIcon"},
-{ "id": "CollectAchievementIcon_s"},
-{ "id": "CultivateAchievementIcon"},
-{ "id": "CultivateAchievementIcon_s"},
-{ "id": "ExploreAchievementIcon"},
-{ "id": "ExploreAchievementIcon_s"},
-{ "id": "IconDecoChallenge"},
-{ "id": "MainLineAchievementIcon"},
-{ "id": "MainLineAchievementIcon_s"},
-{ "id": "ReputationAchievementIcon"},
-{ "id": "ReputationAchievementIcon_s"},
-{ "id": "RougeAchievementIcon"},
-{ "id": "RougeAchievementIcon_s"},
-{ "id": "TrackAchievementIcon"},
-{ "id": "TrackAchievementIcon_s"}
+{ "id": "1101-avatarcutinfigures"},
+{ "id": "1101-avatarcutinbg"},
+{ "id": "1102-avatarcutinfigures"},
+{ "id": "1102B"},
+{ "id": "1103-avatarcutinbg"},
+{ "id": "1103-battledetailicon"},
+{ "id": "1203-avatarcutinbg"},
+{ "id": "1205-avatarcutinfigures"},
+{ "id": "1206-avatarimgfront"},
+{ "id": "1205B"},
+{ "id": "IconRelic_104_2"},
+{ "id": "IconRelic_101_5"},
+{ "id": "IconRelic_105_1"},
+{ "id": "IconRelic_105_3"},
+{ "id": "IconRelic_109_5"},
+{ "id": "IconRelic_111_1"},
+{ "id": "IconRelic_110_6"},
+{ "id": "IconRelic_112_5"},
+{ "id": "IconRelic_111_4"},
+{ "id": "SkillIcon_1004_Ultra"},
+{ "id": "SkillIcon_1013_SkillTree3"},
+{ "id": "SkillIcon_1102_Rank2"}
 ]
